Add `Parser.pre_load` as an extension point
`Parser.pre_load` is a "do nothing" method out of the box, but lets
users subclass and override with custom implementations for their
needs. This resolves #234 by allowing a custom parser to strip
whitespace -- or perform any other desirable action -- after location
loading and before schema loading.

By giving the pre_load method an interface which takes the schema and
location, users can extend this in a myriad of ways.

In an explicit nod towards #234, an example is provided which strips
whitespace from certain locations. Furthermore, since satisfying this
use case is the purpose of this PR, I've included a somewhat unusual
test which copies the doc example code and ensures it does the right
thing in a couple of basic scenarios.

Defining this interface as a method for subclasses to override keeps
the changes to `webargs` minimal (vs. supporting another hook with a
decorator, like error handlers and location loaders).